Support default block in SD LLVM config file
To better support the use case when some modules use SD LLVM while SDCLANG is
set to false globaly, a default block is required in the SD LLVM config file.
SDCLANG_PATH has to be set in the default block so that even if SDCLANG is
false, some modules can still use SD LLVM if needed.